+++ benchmark/hmac_sha1.dart
@@ -33,14 +33,14 @@ class CryptoBenchmark extends Benchmark {
 void main() {
   print('------- HMAC(SHA-1) --------');
   final conditions = [
+    [5 << 20, 10],
+    [1 << 10, 5000],
     [10, 100000],
-    [1000, 5000],
-    [500000, 10],
   ];
   for (var condition in conditions) {
     int size = condition[0];
     int iter = condition[1];
-    print('---- size=$size | iterations: $iter ----');
+    print('---- size: ${formatSize(size)} | iterations: $iter ----');
     HashlibBenchmark(size, iter).showDiff([
       CryptoBenchmark(size, iter),
     ]);