+++ benchmark/pbkdf2.dart
@@ -67,7 +67,7 @@ void main() {
   for (var condition in conditions) {
     int size = condition[0];
     int iter = condition[1];
-    print('---- size=$size | iterations: $iter ----');
+    print('---- size: ${formatSize(size)} | iterations: $iter ----');
     HashlibBenchmark(size, iter).showDiff([
       PointyCastleBenchmark(size, iter),
     ]);