+++ benchmark/ripemd320.dart
@@ -37,14 +37,14 @@ class PointyCastleBenchmark extends Benchmark {
 void main() {
   print('--------- RIPEMD-320 ----------');
   final conditions = [
+    [5 << 20, 10],
+    [1 << 10, 5000],
     [10, 100000],
-    [1000, 5000],
-    [500000, 10],
   ];
   for (var condition in conditions) {
     int size = condition[0];
     int iter = condition[1];
-    print('---- size=$size | iterations: $iter ----');
+    print('---- size: ${formatSize(size)} | iterations: $iter ----');
     HashlibBenchmark(size, iter).showDiff([
       PointyCastleBenchmark(size, iter),
     ]);