+++ scripts/benchmark.sh
@@ -5,5 +5,4 @@ mkdir -p build
 dart compile exe ./benchmark/benchmark.dart -o ./build/benchmark
 chmod +x ./build/benchmark
 ./build/benchmark BENCHMARK.md
-echo ">" >> BENCHMARK.md
-echo "> $(dart --version)" >> BENCHMARK.md
+printf ">\n> $(dart --version)" >> BENCHMARK.md